a little improvment

+++ Sales_item.h
@@ -2,5 +2,6 @@ class Sales_item{
 public:
 	std::string ISBN;
 	int NumbersofSale;
-	float price;
+	double price;
+	//一般使用double而不使用float，因为float会有精度损失，而double引入双精度却没有带来什么性能损失
 };
